Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@762116 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/math/analysis/polynomials/PolynomialFunction.java
@@ -234,7 +234,7 @@ public class PolynomialFunction implements DifferentiableUnivariateRealFunction,
         }
         double[] result = new double[n - 1];
         for (int i = n - 1; i  > 0; i--) {
-            result[i - 1] = (double) i * coefficients[i];
+            result[i - 1] = i * coefficients[i];
         }
         return result;
     }
