+++ src/java/org/apache/commons/math/util/MathUtils.java
@@ -307,7 +307,7 @@ public final class MathUtils {
             return 0;
         }
         if ((k == 1) || (k == n - 1)) {
-            return Math.log((double) n);
+            return Math.log(n);
         }
         
         /*
@@ -337,12 +337,12 @@ public final class MathUtils {
 
         // n!/(n-k)!
         for (int i = n - k + 1; i <= n; i++) {
-            logSum += Math.log((double) i);
+            logSum += Math.log(i);
         }
 
         // divide by k!
         for (int i = 2; i <= k; i++) {
-            logSum -= Math.log((double) i);
+            logSum -= Math.log(i);
         }
 
         return logSum;      
@@ -523,7 +523,7 @@ public final class MathUtils {
         }
         double logSum = 0;
         for (int i = 2; i <= n; i++) {
-            logSum += Math.log((double)i);
+            logSum += Math.log(i);
         }
         return logSum;
     }
